requirements.txt
asgiref==3.8.1
attrs==24.3.0
autopep8==2.3.2
aws-cdk-lib==2.178.2
aws-cdk.asset-awscli-v1==2.2.223
aws-cdk.asset-kubectl-v20==2.1.4
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==39.2.20
babel==2.17.0
cattrs==24.1.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.2.0
constructs==10.4.2
contourpy==1.1.0
cryptography==44.0.2
cycler==0.11.0
Django==4.2.20
django-extra-views==0.14.0
django-haystack==3.3.0
django-localflavor==3.1
django-oscar==3.2.6
django-oscar-paypal @ git+https://github.com/django-oscar/django-oscar-paypal.git@04b77bdf1bdfb5a7123166a7ddeade6b08f0704b
django-phonenumber-field==6.4.0
django-silk==5.3.2
django-tables2==2.7.0
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
factory-boy==3.2.1
Faker==37.1.0
fonttools==4.42.1
gprof2dot==2025.4.14
idna==3.4
importlib_resources==6.5.2
inflection==0.5.1
iniconfig==2.0.0
jsii==1.106.0
kiwisolver==1.4.5
matplotlib==3.7.3
numpy==1.25.2
packaging==23.1
pandas==2.1.0
paypal-checkout-serversdk==1.0.3
paypalhttp==1.0.1
phonenumbers==9.0.3
Pillow==10.0.0
pluggy==1.3.0
publication==0.0.3
purl==1.6
pycodestyle==2.13.0
pycparser==2.22
pygame==2.6.0
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyparsing==3.1.1
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.1.0
python-stdnum==1.20
pytz==2023.3.post1
PyYAML==6.0.2
requests==2.31.0
six==1.16.0
sorl-thumbnail==12.11.0
sqlparse==0.5.3
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5
Whoosh==2.7.4
